# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation Api Gutter Runner",
        "description": "Implémentation de Swagger",
        "contact": {
            "email": "admin@app.gutterrunner.ca"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Serveur d'API de démonstration"
        }
    ],
    "paths": {
...
        "/api/permissions": {
            "get": {
                "tags": [
                    "Roles et Permissions"
                ],
                "summary": "Liste des permissions",
                "description": "Liste des permissions",
                "operationId": "permissionsList",
                "responses": {
                    "200": {
                        "description": "Liste obtenue",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "Uuid de la permission",
                                                        "type": "string",
                                                        "example": "947027"
                                                    },
                                                    "name": {
                                                        "description": "Nom de la permission",
                                                        "type": "string",
                                                        "example": "Create clients"
                                                    },
                                                    "created_at": {
                                                        "description": "Date de création de la permission",
                                                        "type": "string",
                                                        "example": "2021-05-15T17:08:58+00:00"
                                                    },
                                                    "updated_at": {
                                                        "description": "Date de mise à jour de la permission",
                                                        "type": "string",
                                                        "example": "2021-05-15T17:08:58+00:00"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "description": "Métadonnées",
                                            "properties": {
                                                "pagination": {
                                                    "description": "Description de la pagination",
                                                    "properties": {
                                                        "total": {
                                                            "description": "Total des données",
                                                            "type": "integer",
                                                            "example": 13
                                                        },
                                                        "count": {
                                                            "description": "Nombre de données affichées",
                                                            "type": "integer",
                                                            "example": 13
                                                        },
                                                        "per_page": {
                                                            "description": "Nombre de données affichées par page",
                                                            "type": "integer",
                                                            "example": 15
                                                        },
                                                        "current_page": {
                                                            "description": "Numéro de la page courante",
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "total_pages": {
                                                            "description": "Nombre total des pages",
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "links": {
                                                            "description": "Liens associés",
                                                            "type": "object"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Utilisateur non connecté",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "L'utilisateur n'est même pas connecté",
                                            "type": "string",
                                            "example": "Vous ne vous êtes pas encore authentifié."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Utilisateur non autorisé ou non permis",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "L'utilisateur n'a pas la permission nécessaire",
                                            "type": "string",
                                            "example": "Vous n'avez pas la permission d'effectuer cette action."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer token",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}